requirements.txt
fastapi
uvicorn
pandas
beautifulsoup4
requests
jinja2
sendgrid
python-dotenv
python-multipart

